Plz Read Description
The map was never going to work so tried to fix it. When doing a conversion mod you need to learn to walk before you can run instead of trying to do it in one go. The height and terrain map survived but provinces were reverted back to vanilla. All of the old stuff is still around in the old mod folder in case its needed for any reason.